test/K8sServices: fix a call to helpers.CurlFail()
helpers.CurlFail() can take optional arguments, and at one location this
is used to pass the "--max-time" option to the curl command line.

But these optional arguments for helpers.CurlFail() are not intended to
be used that way; they should instead be used to format the target URL,
which is supposed to be a string containing the relevant format
specifiers. Because in the current case, the URL provided has no format
specifier, go appends the additional arguments at the end, with
extraneous characters. So instead of producing the final curl command:

    curl [...] http://192.168.36.11:30861 --max-time 3 [...]

We actually get:

    curl [...] --max-time 20 http://192.168.36.11:30861%!(EXTRA string=--max-time 3) [...]

The first "--max-time 20" is added by helpers.CurlFail() itself.

Then of course bash fails to interpret that command, and complains with:

    bash: -c: line 0: syntax error near unexpected token `('

In order to fix it, let's remove the "--max-time 3" argument entirely.
As we can see from the resulting command line, "--max-time" is already
used (with "CurlMaxTimeout"). And it is useless anyway, since we
actually expect to timeout on the value passed through the
"--connect-timeout" option (currently 5 seconds).

While at it, let's also fix the typo in the comments for
helpers.CurlFail(), and remove the explicit argument indices used for
generating the curl command. They were useful at a time when one of the
arguments was used multiple times, but they are unnecessary and only
confusing now.

Fixes: 54304819a1d2 ("test: Extend externalTrafficPolicy=Local tests")
Signed-off-by: Quentin Monnet <quentin@isovalent.com>