test: enquote bugtool archive folder name in bash command
There are a couple of tests that contain parenthesis in their names.
Namely, in K8sServices:

- Tests NodePort (kube-proxy)
- Tests NodePort (kube-proxy) with externalTrafficPolicy=Local

Because of the parenthesis, handling directories named after the tests
can be more delicate in bash, and we need to make sure that the
parenthesis are not interpreted. So far this was not done properly, and
bash complains with:

    bash: -c: line 0: syntax error near unexpected token `('

Fix it by enquoting the folder name in the relevant bash commands.

Fixes: #11094
Signed-off-by: Quentin Monnet <quentin@isovalent.com>